+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Add
 - Added wheel package supported for `py3.12`
+- Added `figsize` param in `toad.plot.bin_plot` function
 
 ### Changed
 - Update `pandas` version to `>=1.5`